requirements.txt

PyQt5>=5.15.0
paho-mqtt>=1.6.0
amqtt>=0.11.0
zeroconf>=0.131.0
requests>=2.32.0
PyYAML>=6.0
Flask>=3.0.0
Pillow>=10.0.0
qrcode>=7.0
pywin32>=306; sys_platform == "win32"
pyinstaller>=6.0.0